Removing testDevices from appleauth
