+++ appleauth/fetch.go
@@ -2,7 +2,6 @@ package appleauth
 
 import (
 	"fmt"
-	"net/http"
 
 	"github.com/bitrise-io/go-utils/log"
 	"github.com/bitrise-steplib/steps-deploy-to-itunesconnect-deliver/devportalservice"
@@ -10,9 +9,8 @@ import (
 
 // Credentials contains either Apple ID or APIKey auth info
 type Credentials struct {
-	AppleID     *AppleID
-	APIKey      *devportalservice.JWTConnection
-	TestDevices []devportalservice.TestDevice
+	AppleID *AppleID
+	APIKey  *devportalservice.JWTConnection
 }
 
 // AppleID contains Apple ID auth info
@@ -39,34 +37,11 @@ func (*MissingAuthConfigError) Error() string {
 	return "Apple Service authentication not configured"
 }
 
-const notConnected = `Connected Apple Developer Portal Account not found.
-Most likely because there is no Apple Developer Portal Account connected to the build, or the build is running locally.
-Read more: https://devcenter.bitrise.io/getting-started/configuring-bitrise-steps-that-require-apple-developer-account-data/`
-
 // Select return valid Apple ID or API Key based authentication data, from the provided Bitrise Apple Developer Connection or Inputs
 // authSources: required, array of checked sources (in order, the first set one will be used)
 //	 for example: []AppleAuthSource{&SourceConnectionAPIKey{}, &SourceConnectionAppleID{}, &SourceInputAPIKey{}, &SourceInputAppleID{}}
 // inputs: optional, user provided inputs that are not centrally managed (by setting up connections)
-func Select(devportalConnectionProvider devportalservice.AppleDeveloperConnectionProvider, authSources []Source, inputs Inputs) (Credentials, error) {
-	initializeConnection := false
-	for _, source := range authSources {
-		initializeConnection = initializeConnection || source.RequiresConnection()
-	}
-
-	var conn *devportalservice.AppleDeveloperConnection
-	if initializeConnection && devportalConnectionProvider != nil {
-		var err error
-		conn, err = devportalConnectionProvider.GetAppleDeveloperConnection()
-		if err != nil {
-			handleSessionDataError(err)
-		}
-
-		if conn == nil || (conn.JWTConnection == nil && conn.SessionConnection == nil) {
-			fmt.Println()
-			log.Debugf("%s", notConnected)
-		}
-	}
-
+func Select(conn *devportalservice.AppleDeveloperConnection, authSources []Source, inputs Inputs) (Credentials, error) {
 	for _, source := range authSources {
 		auth, err := source.Fetch(conn, inputs)
 		if err != nil {
@@ -82,19 +57,4 @@ func Select(devportalConnectionProvider devportalservice.AppleDeveloperConnectio
 	}
 
 	return Credentials{}, &MissingAuthConfigError{}
-}
-
-func handleSessionDataError(err error) {
-	if err == nil {
-		return
-	}
-
-	if networkErr, ok := err.(devportalservice.NetworkError); ok && networkErr.Status == http.StatusNotFound {
-		log.Debugf("")
-		log.Debugf("%s", notConnected)
-	} else {
-		fmt.Println()
-		log.Errorf("Failed to activate Bitrise Apple Developer Portal connection: %s", err)
-		log.Warnf("Read more: https://devcenter.bitrise.io/getting-started/configuring-bitrise-steps-that-require-apple-developer-account-data/")
-	}
 }
