+++ appleauth/fetch_test.go
@@ -10,9 +10,9 @@ import (
 
 func TestSelect(t *testing.T) {
 	type args struct {
-		devportalConnectionProvider devportalservice.AppleDeveloperConnectionProvider
-		authSources                 []Source
-		inputs                      Inputs
+		devportalConnection *devportalservice.AppleDeveloperConnection
+		authSources         []Source
+		inputs              Inputs
 	}
 	tests := []struct {
 		name        string
@@ -24,9 +24,9 @@ func TestSelect(t *testing.T) {
 		{
 			name: "No connection active (nil), no inputs",
 			args: args{
-				devportalConnectionProvider: newMockDevportalConnectionProvider(nil, nil),
-				authSources:                 []Source{&ConnectionAPIKeySource{}, &ConnectionAppleIDSource{}, &InputAPIKeySource{}, &InputAppleIDSource{}},
-				inputs:                      Inputs{},
+				devportalConnection: nil,
+				authSources:         []Source{&ConnectionAPIKeySource{}, &ConnectionAppleIDSource{}, &InputAPIKeySource{}, &InputAppleIDSource{}},
+				inputs:              Inputs{},
 			},
 			want:        Credentials{},
 			wantErr:     true,
@@ -35,20 +35,9 @@ func TestSelect(t *testing.T) {
 		{
 			name: "No connection active (empty), no inputs",
 			args: args{
-				devportalConnectionProvider: newMockDevportalConnectionProvider(&devportalservice.AppleDeveloperConnection{}, nil),
-				authSources:                 []Source{&ConnectionAPIKeySource{}, &ConnectionAppleIDSource{}, &InputAPIKeySource{}, &InputAppleIDSource{}},
-				inputs:                      Inputs{},
-			},
-			want:        Credentials{},
-			wantErr:     true,
-			wantErrType: &MissingAuthConfigError{},
-		},
-		{
-			name: "No connection active (empty, error), no inputs",
-			args: args{
-				devportalConnectionProvider: newMockDevportalConnectionProvider(&devportalservice.AppleDeveloperConnection{}, devportalservice.NetworkError{}),
-				authSources:                 []Source{&ConnectionAPIKeySource{}, &ConnectionAppleIDSource{}, &InputAPIKeySource{}, &InputAppleIDSource{}},
-				inputs:                      Inputs{},
+				devportalConnection: &devportalservice.AppleDeveloperConnection{},
+				authSources:         []Source{&ConnectionAPIKeySource{}, &ConnectionAppleIDSource{}, &InputAPIKeySource{}, &InputAppleIDSource{}},
+				inputs:              Inputs{},
 			},
 			want:        Credentials{},
 			wantErr:     true,
@@ -57,8 +46,8 @@ func TestSelect(t *testing.T) {
 		{
 			name: "No connection active (empty, error), inputs (Apple ID)",
 			args: args{
-				devportalConnectionProvider: newMockDevportalConnectionProvider(&devportalservice.AppleDeveloperConnection{}, nil),
-				authSources:                 []Source{&ConnectionAPIKeySource{}, &ConnectionAppleIDSource{}, &InputAPIKeySource{}, &InputAppleIDSource{}},
+				devportalConnection: &devportalservice.AppleDeveloperConnection{},
+				authSources:         []Source{&ConnectionAPIKeySource{}, &ConnectionAppleIDSource{}, &InputAPIKeySource{}, &InputAppleIDSource{}},
 				inputs: Inputs{
 					Username: "a", Password: "b", AppSpecificPassword: "c",
 					APIIssuer: "", APIKeyPath: "",
@@ -74,11 +63,11 @@ func TestSelect(t *testing.T) {
 		{
 			name: "Connection active (API Key), inputs (Apple ID)",
 			args: args{
-				devportalConnectionProvider: newMockDevportalConnectionProvider(&devportalservice.AppleDeveloperConnection{
+				devportalConnection: &devportalservice.AppleDeveloperConnection{
 					JWTConnection: &devportalservice.JWTConnection{
 						KeyID: "x", IssuerID: "y", PrivateKey: "z",
 					},
-				}, nil),
+				},
 				authSources: []Source{&ConnectionAPIKeySource{}, &ConnectionAppleIDSource{}, &InputAPIKeySource{}, &InputAppleIDSource{}},
 				inputs: Inputs{
 					Username: "a", Password: "b", AppSpecificPassword: "c",
@@ -95,11 +84,11 @@ func TestSelect(t *testing.T) {
 		{
 			name: "Connection active (API Key), inputs (Apple ID), connection not enabled",
 			args: args{
-				devportalConnectionProvider: newMockDevportalConnectionProvider(&devportalservice.AppleDeveloperConnection{
+				devportalConnection: &devportalservice.AppleDeveloperConnection{
 					JWTConnection: &devportalservice.JWTConnection{
 						KeyID: "x", IssuerID: "y", PrivateKey: "z",
 					},
-				}, nil),
+				},
 				authSources: []Source{&InputAPIKeySource{}, &InputAppleIDSource{}},
 				inputs: Inputs{
 					Username: "a", Password: "b", AppSpecificPassword: "c",
@@ -116,7 +105,7 @@ func TestSelect(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := Select(tt.args.devportalConnectionProvider, tt.args.authSources, tt.args.inputs)
+			got, err := Select(tt.args.devportalConnection, tt.args.authSources, tt.args.inputs)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("Select() error = %v, wantErr %v", err, tt.wantErr)
 				return
@@ -125,17 +114,4 @@ func TestSelect(t *testing.T) {
 			require.Equal(t, tt.want, got, "Select() =")
 		})
 	}
-}
-
-type mockDevportalConnectionProvider struct {
-	conn *devportalservice.AppleDeveloperConnection
-	err  error
-}
-
-func (m *mockDevportalConnectionProvider) GetAppleDeveloperConnection() (*devportalservice.AppleDeveloperConnection, error) {
-	return m.conn, m.err
-}
-
-func newMockDevportalConnectionProvider(conn *devportalservice.AppleDeveloperConnection, err error) devportalservice.AppleDeveloperConnectionProvider {
-	return &mockDevportalConnectionProvider{conn: conn, err: err}
 }
